Fixing tests.

+++ models/dtos/order/order_read_dto.py
@@ -4,6 +4,7 @@ from typing import List, Optional
 from pydantic import BaseModel
 from models.enums import OrderStatus
 from models.dtos.order_item.order_item_read_dto import OrderItemReadDTO
+from decimal import Decimal
 
 
 class OrderReadDTO(BaseModel):
@@ -12,7 +13,7 @@ class OrderReadDTO(BaseModel):
     updated_at: Optional[datetime] = None
     deleted_at: Optional[datetime] = None
     customer_id: uuid.UUID
-    total_amount: float
+    total_amount: Decimal
     status: OrderStatus
     items: List[OrderItemReadDTO] = []
